Add scripts and XMLs.
Most of the figures should be easy to replicate by using the IPython
notebook and the data that's already in the repository. I also added
the XMLs to run homoplasy and marginal likelihood analyses in BEAST.
Because of the file sizes associated with Bayesian phylogenetics I
don't think I can put the posterior sets of trees in the repository.
The XMLs included will allow to run those analyses, however.